Rename spec files, remove mainnet & testnet directories
