+++ package.json
@@ -18,8 +18,8 @@
     "sync-versions": "node script/sync-versions.js",
     "test:unit": "jest tests/unit",
     "postinstall": "husky install",
-    "test:testnet": "node_modules/mocha/bin/mocha --colors 'tests/**/testnet/*.spec.js'",
-    "test:mainnet": "NODE_ENV=mainnet node_modules/mocha/bin/mocha --colors --grep 'mainnet' 'tests/**/testnet/*.spec.js'"
+    "test:testnet": "node_modules/mocha/bin/mocha --colors 'tests/**/*.spec.js'",
+    "test:mainnet": "NODE_ENV=mainnet node_modules/mocha/bin/mocha --colors --grep 'mainnet' 'tests/**/*.spec.js'"
   },
   "dependencies": {
     "@liquality/bitcoin-esplora-batch-api-provider": "^1.1.9",
