+++ tests/Pages/ReceivePage.js
