+++ tests/integration/send_testnet.spec.js
@@ -1,11 +1,11 @@
-const TestUtil = require('../../utils/TestUtils')
+const TestUtil = require('../utils/TestUtils')
 const OverviewPage = require('../Pages/OverviewPage')
 const HomePage = require('../Pages/HomePage')
 const PasswordPage = require('../Pages/PasswordPage')
 const SearchAssetPage = require('../Pages/SearchAssetPage')
 const SendPage = require('../Pages/SendPage')
 const TransactionDetailsPage = require('../Pages/TransactionDetailsPage')
-const TestDataUtils = require('../../utils/TestDataUtils')
+const TestDataUtils = require('../utils/TestDataUtils')
 const expect = require('chai').expect
 const chalk = require('chalk')
